Refactoring admin routes out of the Products file
